+++ docs/changelog.md
@@ -4,6 +4,8 @@
 
 ### 2024
 
+- April/10; **0.1.4**: Playlist header and tracklist processing + parsing moved to server from client.
+
 - **0.1.3**: Full tracklist fetching & more robust tracklist type conversion
 
 - **0.1.2**: Tracklist received more customization